Move Ship Editor Files And Coverity Clean-up

+++ qtfred/src/ui/dialogs/ShipEditor/PlayerOrdersDialog.cpp
@@ -20,7 +20,7 @@ namespace fso {
 				ui->setupUi(this);
 				connect(this, &QDialog::accepted, _model.get(), &PlayerOrdersDialogModel::apply);
 				connect(this, &QDialog::rejected, _model.get(), &PlayerOrdersDialogModel::reject);
-				for (int i = 0; i < (int)_model->getAcceptedOrders().size(); i++) {
+				for (size_t i = 0; i < _model->getAcceptedOrders().size(); i++) {
 					//i == 0 check added to avoid culling first entry where getAcceptedOrders returns 0
 					if (_model->getAcceptedOrders()[i] || i == 0) {
 						check_boxes.push_back(new ShipFlagCheckbox(nullptr));
@@ -71,7 +71,7 @@ namespace fso {
 			void PlayerOrdersDialog::updateUI()
 			{
 				util::SignalBlockers blockers(this);
-				for (int i = 0; i < (int)check_boxes.size(); i++) {
+				for (size_t i = 0; i < check_boxes.size(); i++) {
 					int state = _model->getCurrentOrders()[i];
 					switch (state) {
 					case 0:
